Check if user has book and added model for reading progress
Added a route for the book-content, where the user can search for a book's content using the bookuuid, NOT THE book-content id, and check to see if the user has this book, if not it returns an error saying that the user does not have the book, if the user is not logged in it also returns a message telling the user he has to be logged in. This is to ensure that not anyone can access the book's content unless they have it in their library.

Added a User-book-progress model which has a many to many relation with both the User model and the Book model, with an extra attribute which has the page number the user has stopped at reading